VideoPress: rename Phase-4 component files to kebab-case
The codebase already has a mix of naming styles for component files
(`Library/upload-actions-context.tsx` vs `Library/ThumbnailField.tsx`).
Standardising on kebab-case for files going forward; this commit
covers the six Video-details components added in Phase 4. The
`VideoDetails/` directory keeps its PascalCase name to match the
existing `Library/` / `DashboardLayout/` precedent — file-name-only
change.
