+++ projects/packages/videopress/routes/video/stage.tsx
@@ -5,13 +5,13 @@ import { useCallback, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { Link, useNavigate, useParams } from '@wordpress/route';
 import { Stack, Text } from '@wordpress/ui';
-import ChaptersHelpModal from '../../src/dashboard/components/VideoDetails/ChaptersHelpModal';
-import HeaderActions from '../../src/dashboard/components/VideoDetails/HeaderActions';
-import PrivacySharingCard from '../../src/dashboard/components/VideoDetails/PrivacySharingCard';
-import RatingCard from '../../src/dashboard/components/VideoDetails/RatingCard';
-import ThumbnailCard from '../../src/dashboard/components/VideoDetails/ThumbnailCard';
-import VideoDetailsCard from '../../src/dashboard/components/VideoDetails/VideoDetailsCard';
+import ChaptersHelpModal from '../../src/dashboard/components/VideoDetails/chapters-help-modal';
+import HeaderActions from '../../src/dashboard/components/VideoDetails/header-actions';
+import PrivacySharingCard from '../../src/dashboard/components/VideoDetails/privacy-sharing-card';
+import RatingCard from '../../src/dashboard/components/VideoDetails/rating-card';
+import ThumbnailCard from '../../src/dashboard/components/VideoDetails/thumbnail-card';
 import { useVideoDetailsForm } from '../../src/dashboard/components/VideoDetails/use-video-details-form';
+import VideoDetailsCard from '../../src/dashboard/components/VideoDetails/video-details-card';
 import { useMockLibrary } from '../../src/dashboard/hooks/use-mock-library';
 import './style.scss';
 import type { MockLibraryItem } from '../../src/dashboard/types/library';
