+++ projects/packages/videopress/src/dashboard/components/VideoDetails/rating-card.tsx
