+++ projects/packages/videopress/src/dashboard/components/VideoDetails/thumbnail-card.tsx
